Make `JsonWebKey` a `partial dictionary`
To ease automatic processing of IDL defined in this spec and in Web Crypto, it
would be great if, as much as possible, this spec "extended" the constructs it
needs to extend from Web Crypto instead of re-defining them.

The spec re-defined the `JsonWebKey` dictionary. This update turns that
re-definition into an extension, using `partial dictionary`, as done for
`SubtleCrypto`.

Note: The same approach unfortunately cannot work for `KeyFormat` and
`KeyUsage` as Web IDL does not support partial enumerations.