Make `JsonWebKey` a `partial dictionary` (#26)
To ease automatic processing of IDL defined in this spec and in Web Crypto,
it's best if, as much as possible, this spec "extends" the constructs it
needs to extend from Web Crypto instead of re-defining them.

The spec previously re-defined the `JsonWebKey` dictionary. This update
turns that re-definition into an extension, using `partial dictionary`,
as done for `SubtleCrypto`.

Note: The same approach unfortunately cannot work for `KeyFormat` and
`KeyUsage` as Web IDL does not support partial enumerations.